Implemented HSV color picking, using PWM module

+++ esl_utils.c
@@ -1,6 +1,6 @@
 #include "esl_utils.h"
 
-void hsv_to_rgb(uint32_t hue, uint32_t saturation, uint32_t value, uint32_t *r, uint32_t *g, uint32_t *b ) {
+void hsv_to_rgb(uint16_t hue, uint8_t saturation, uint8_t value, uint8_t *r, uint8_t *g, uint8_t *b ) {
     float h = hue / 60.0;  // Sector of 60 degrees
     float s = saturation / 100.0;
     float v = value / 100.0;
